+++ js/widgets/appbar.js
@@ -371,16 +371,18 @@
                     //===  EVENTS =================================================
 
                     //activate the click event for the pull button
-                    $(that.pullButton).on("click", function () {
+                    $(that.pullButton).on("click", function (e) {
 
                         //who am i?
                         that = $(this).closest("[data-role=appbar]").data("appbar");
 
                         //we show /hide the pullmenu
-                        if ($(that.pullMenu).is(":hidden")) {
+                        if ($(that.pullMenu).length !== 0 && $(that.pullMenu).is(":hidden")) {
                             $(that.pullMenu).show();
                             $(that.pullMenu).find(".app-bar-pullmenubar")
                                     .hide().not(".hidden").slideDown("fast");
+                            e.preventDefault();
+                            e.stopPropagation();
                         } else {
                             $(that.pullMenu).find(".app-bar-pullmenubar")
                                     .not(".hidden").show().slideUp("fast", function () {