+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog 
 
+## 3.0.18
+* add method isOpened for Dialog widget 
+* add method metroDialog.isDialog(el) 
+
 ## 3.0.17
 * fix hotkeys init error